requirements.txt
weaviate-client>=0.5.0
wheel
twine
django>=2.2
pytest>=5.0.0
pytest-mock>=1.0.0 
mkdocs>=1.0.0
mkdocs-material>=3.0.0